requirements.txt
customtkinter>=5.2.0
pyscard>=2.0.7
pycryptodome>=3.20.0
Pillow>=10.0.0
